refactor(blue) : 사다리 생성 전략 패키지 생성
