+++ src/main/java/domain/bridgeConstructstrategy/BridgeConstructStrategy.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.bridgeConstructstrategy;
+
+import domain.Bridges;
 
 public interface BridgeConstructStrategy {
     Bridges generate(int count);
